strstr, strlcat

+++ lib/ft_strsub.c
@@ -9,7 +9,7 @@ char    *ft_strsub(char const *s, unsigned int start, size_t len)
         return (NULL);
     t = str;
     s += start;
-    while (t - str < len)
+    while ((size_t)(t - str) < len)
         *t++ = *s++;
     *t = '\0';
     return (str);
